Move external utils file to io package.
