+++ src/main/java/fi/helsinki/cs/tmc/cli/updater/TmcCliUpdater.java
@@ -1,7 +1,7 @@
 package fi.helsinki.cs.tmc.cli.updater;
 
+import fi.helsinki.cs.tmc.cli.io.ExternalsUtil;
 import fi.helsinki.cs.tmc.cli.io.Io;
-import fi.helsinki.cs.tmc.cli.tmcstuff.ExternalsUtil;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
